commit by rahul regression points 23-09-2022

+++ app/src/main/java/com/agent/cashmoovui/transactionhistory_walletscreen/TransactionListLisners.java
@@ -1,7 +1,7 @@
 package com.agent.cashmoovui.transactionhistory_walletscreen;
 
 public interface TransactionListLisners {
-    void onTransactionViewItemClick(String transId, String transType, String transDate, String source, String destination, int sourceMsisdn, int destMsisdn, String symbol, int amount, int fee, String taxType, String tax, int postBalance, String status);
+    void onTransactionViewItemClick(String transId, String transType, String transDate, String source, String destination, int sourceMsisdn, int destMsisdn, String symbol, int amount, int fee, String taxType, String tax, double postBalance, String status);
 
 }
